Allow vagrant user NOPASSWD sudo to all accounts
Add an 'ALL' Runas_Spec to the sudoers.d/vagrant entry to allow the
vagrant account to run any command as any user, rather than just as the
root user.

Having amended this entry, there is no longer a need to add the vagrant
account to the sudo group which allows execution of any command as any
user WITH a password.